Leaderboard implementation with SQL
Upon application start the leaderboard initializes an SQL database. Whenever a game is finished (win/lose) the GameManager orders to save the score, Afterwards the leaderboard window is opened and the last game is highlighted (When opened from homescreen nothing is highlighted)

The leaderboardWindow GameObject has to be active when the application is loading got initialization.

There is only one instance of this window throughout all scenses (DontDestroyOnLoad) and implements a Singleton pattern.

Leaderboard GUI is a table containing a TableHeader and a body that parents all rows (prefab)

Added GlobalVariables static class which holds the username and the selected configuration to be accessed from everywhere. AppConfig class moved to the same file aswell (Misc.cs). PlayerName is saved and stored on PlayerPrefs

#24
closes #10
closes #21