Updated ConfigLoader to return a list (even more generic!) and added explanations for it.
Using a list instead of a dictionary because it can be mapped later. (Can't really deserialize a list to a dictionary though..)

Decoupled the ParseRawData method and the Deserialize method to allow addition of new config types by adding just Deserialize (Parsing process should be the same for all type). The code is a lot cleaner now.

Updated config files to include a name and the sereialized classes to match them.
UnitTests were updated accordingly.
(It seems that the cost of generalizing it costs in runtime- its much slower)

Updated Unity project to use a list for configs, there is no need for a dictionary. Could now also load a list of configurations (and not just an object)

Added explanations to ConfigLoader.
#22